requirements.txt
# Schwab API
schwab-py>=1.5.1
python-dotenv>=1.0.0

# Web Interface
Flask>=3.0.0
markdown>=3.5.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Statistics / Black-Scholes
scipy>=1.11.0

# Optional – nicer printing (not required for logic)
tabulate>=0.9.0
